LayoutTests/editing/spelling/context-menu-suggestions-multiword-selection-expected.txt

Spell check sugestions should not be pupulated when user selects multiple words that are not a single misspelling. To test manually, type 'wellcome home.', select 'wellcome home', and context-click on the selection. The test suceeds when the context menu has no spell check suggestions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is ""
PASS window.getSelection().toString() is "wellcome home"
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is "wellcome home"
PASS contextMenuItemsWithoutSpellingSuggestions < contextMenuItemsWithSpellingSuggestions is true
PASS successfullyParsed is true

TEST COMPLETE
wellcome home.
